+++ compositor_pipeline/src/pipeline/webrtc/server/create_whep_session.rs
@@ -1,8 +1,10 @@
 use crate::pipeline::webrtc::{
-    error::WhipWhepServerError, handle_keyframe_requests::handle_keyframe_requests,
-    handle_packet_loss_requests::handle_packet_loss_requests,
-    peer_connection_sendonly::SendonlyPeerConnection,
-    whep_output::stream_media_to_peer::spawn_media_streaming_task, WhipWhepServerState,
+    error::WhipWhepServerError,
+    handle_keyframe_requests::handle_keyframe_requests,
+    whep_output::{
+        peer_connection::PeerConnection, stream_media_to_peer::spawn_media_streaming_task,
+    },
+    WhipWhepServerState,
 };
 use axum::{
     body::Body,
@@ -32,7 +34,7 @@ pub async fn handle_create_whep_session(
     let video_options = outputs.get_with(&output_id, |output| Ok(output.video_options.clone()))?;
     let audio_options = outputs.get_with(&output_id, |output| Ok(output.audio_options.clone()))?;
 
-    let peer_connection = SendonlyPeerConnection::new(&ctx.clone()).await?;
+    let peer_connection = PeerConnection::new(&ctx.clone()).await?;
 
     let (video_track, video_sender) = match video_options.clone() {
         Some(opts) => {
@@ -42,12 +44,9 @@ pub async fn handle_create_whep_session(
         None => (None, None),
     };
 
-    let (audio_track, audio_sender) = match audio_options.clone() {
-        Some(opts) => {
-            let (track, sender) = peer_connection.new_audio_track(opts.encoder).await?;
-            (Some(track), Some(sender))
-        }
-        None => (None, None),
+    let audio_track = match audio_options.clone() {
+        Some(opts) => Some(peer_connection.new_audio_track(opts.encoder).await?),
+        None => None,
     };
 
     let sdp_answer = peer_connection.negotiate_connection(offer).await?;
@@ -71,16 +70,6 @@ pub async fn handle_create_whep_session(
                 .track_thread_handle
                 .keyframe_request_sender
                 .clone(),
-        );
-    }
-
-    if let (Some(sender), Some(audio_opt)) = (audio_sender, audio_options) {
-        handle_packet_loss_requests(
-            &ctx,
-            peer_connection.get_rtc_peer_connection(),
-            sender.clone(),
-            audio_opt.track_thread_handle.packet_loss_sender.clone(),
-            audio_opt.ssrc,
         );
     }
