+++ compositor_pipeline/src/pipeline/webrtc/whip_output/peer_connection.rs
@@ -18,6 +18,7 @@ use webrtc::{
         rtp_sender::RTCRtpSender,
         RTCPFeedback,
     },
+    stats::StatsReport,
     track::track_local::track_local_static_rtp::TrackLocalStaticRTP,
 };
 
@@ -137,8 +138,8 @@ impl PeerConnection {
         self.pc.on_ice_candidate(f);
     }
 
-    pub fn get_rtc_peer_connection(&self) -> Arc<RTCPeerConnection> {
-        self.pc.clone()
+    pub async fn get_stats(&self) -> StatsReport {
+        self.pc.get_stats().await
     }
 }
